Update from xip_0.3-src archive -- xip_0.3-xipinv240-vs9-win32-itk3.8.0-vtk5.2.0-release

+++ src/database/remote/losslessCompression/huffmanCoder.cpp
@@ -527,7 +527,7 @@ int huffmanCoder::HUFFMAN_C_Diff( unsigned char *in1, const unsigned int insize,
   sStream stream;
   unsigned int k, total_bytes;
 
-  if (mInSize < insize)
+  if (mInSize < (int)insize)
   {
 	  delete[] diffStream;
 	  diffStream = new unsigned char[insize];
